Change day 13 solution to use linear equations instead of matrices and solve part two

+++ Day13 - Claw Contraption/GlobalUsings.cs
@@ -1,2 +1,2 @@
-﻿global using Point = AdventOfCode.Common.Geometry.Point2D<int>;
-global using Vector = AdventOfCode.Common.Geometry.Vector2D<int>;
+﻿global using Point = AdventOfCode.Common.Geometry.Point2D<long>;
+global using Vector = AdventOfCode.Common.Geometry.Vector2D<long>;
